Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-createAnswer-expected.txt


PASS createAnswer() with null remoteDescription should reject with InvalidStateError
PASS createAnswer() after setting remote description should succeed
PASS createAnswer() when connection is closed reject with InvalidStateError
